moved print functions to reporting module
